feat(cluster-tax): log-domain cluster-factor curve in picocredits (#626 PR 1/3) (#627)
Replace ClusterFactorCurve's linear-domain sigmoid (calibrated in simulator
units, w_mid=10M) with the normative log-domain construction from the #626
spec, consumed in on-chain picocredits. This fixes the step-function collapse
where every live cluster was pinned at ~6x, and de-quantizes the factor output
(the old code snapped to whole multipliers and capped at 5000).

Curve (normative integer algorithm, deterministic, no floats):
- log2_fp(w): (127 - leading_zeros) integer part + top-16 mantissa bits, Q16
- z_fp = log2_fp(W) - LOG2_WMID_FP (const 3_695_429); x = (z_fp*1000)/131_072
  (i64 truncating division is normative); existing 7-knot sigmoid LUT with
  tails changed to exact saturation (x<=-6000 -> 0, x>=6000 -> 65536)
- output = 1000 + (5000 * sig) / 65536; W=0 -> 1000 exactly
- W_MID_PICO = 100_000 * PICO_PER_BTH (=1e17); midpoint = exactly 3.5x
- compile-time locks: W_MID_PICO == 100_000*PICO_PER_BTH, LOG2_WMID_FP==3_695_429

API: ClusterFactorCurve::factor and ZkFeeCurve::factor take u128 picocredit
wealth (was u64). Call sites (mempool/lottery/demurrage/store) pass `w as u128`
for now; the cluster_wealth_db u128 widening is PR 2.

ZkFeeCurve re-anchored to 5 segments at power-of-two multiples of W_MID_PICO
[0, W>>8, W>>3, W<<3, W<<8, u128::MAX], factors [1000,1090,2071,4928,5909,6000]
sampled from the sigmoid; boundaries/SegmentParams widened to u128.

Tests: golden-vector table (0/10 BTH->1000, 100->1050, 1k->1265, 10k->1939,
100k->3500, 1M->5093, 10M->5745, >=409.6M & u128::MAX->6000), monotonicity
(200k random u128 pairs + exhaustive octave-boundary sweep), unit-consistency
guard (50 BTH < 1.1x), testnet baseline (50 BTH=1030, 97.5 BTH=1049), LUT
saturation tails, constant locks, ZkFeeCurve agreement.

Also fixes the nanoBTH->picocredit doc labeling in FeeConfig (same unit-bug
class). Consensus-breaking (changes C7 floor + lottery tilt + mempool fee);
rides the bundled protocol-4.0.0 testnet reset.

Updates #626

+++ botho/src/consensus/lottery.rs
@@ -1296,7 +1296,7 @@ mod tests {
         let curve = ClusterFactorCurve::default_params();
         // Valid mid-range wealth and the u64 extremes all stay in [1000, 6000].
         for w in [0u64, 10_000_000, u64::MAX] {
-            let f = curve.factor(w);
+            let f = curve.factor(w as u128);
             assert!(
                 (FACTOR_SCALE..=MAX_FACTOR_SCALED).contains(&f),
                 "factor {} out of [{}, {}] for wealth {}",
